add modified simplify function
